backend/requirements.txt
asgiref==3.5.2
bcrypt==3.2.2
beanie==1.13.0
cffi==1.15.1
click==8.1.3
cryptography==3.4.8
email-validator==1.3.0
fastapi==0.68.2
fastapi-jwt-auth==0.5.0
fastapi-mail
h11==0.13.0
motor==3.0.0
passlib==1.7.4
pycparser==2.21
pydantic==1.10.2
PyJWT==1.7.1
pymongo==4.1.1
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
six==1.16.0
starlette==0.14.2
typing_extensions==4.3.0
uvicorn==0.15.0
websockets==10.3
yarl==1.8.1

